+++ client/src/components/admin/AdminList/AdminList.module.css
@@ -1,5 +1,4 @@
-.userList{
-    flex: 7;
-    margin-left: 30px;
-    margin-right: 30px;
+.userList {
+    width: 100%;
+    padding: 0 2rem;
 }
