+++ client/src/components/admin/AdminView.tsx
@@ -13,16 +13,16 @@ import DonationsList from "./DonationsList/DonationsList";
 
 
 export default function AdminView() {
-  
+
   // const [actualPage, setActualPage] = useState("home")
   // {actualPage === "home"? <HomeAdmin/> : actualPage === "userList"? <UserList/> : actualPage === "User"? <User /> : null}
 
   return (
-    
+
     <React.Fragment>
-      <TopBar/>
+      <TopBar />
       <div className={st.container}>
-        <SideBar/>
+        <SideBar />
 
         <Routes>
           <Route path='/' element={<HomeAdmin />} />
@@ -32,9 +32,8 @@ export default function AdminView() {
           <Route path='/user/:userId' element={<User />} />
           <Route path='/donations' element={<DonationsList />} />
           {/* <Route path='/donations' element={<Memberships />} /> */}
-          <Route path='*' element={<HomeAdmin />} />
         </Routes>
-          
+
       </div>
     </React.Fragment>
   )
