+++ client/src/components/admin/HelpUsMail/HelpUsMail.module.css
@@ -1,7 +1,6 @@
 .Container {
-    display: flex;
-    flex-direction: column;
     width: 100%;
+    padding-left: 2rem;
 }
 
 .title {
@@ -10,33 +9,18 @@
     margin: 0px;
 }
 
-.listContainer {
+.listMessages {
     display: flex;
-    padding: 0px 10px;
-    text-align: center;
-    margin: 0px auto;
+    flex-wrap: wrap;
+    gap: 1rem;
+    justify-content: center;
+    padding-top: 2rem;
 }
 
 /* .listMessages{
     background-color: blueviolet;
 } */
 
-.divItem {
-    border-radius: 10px;
-    margin: 20px 20px 30px 0px;
-    padding: 5px;
-    -webkit-box-shadow: 0px 0px 15px 2px #000000;
-    box-shadow: 0px 0px 15px 2px #000000;
-    background: hsla(263, 54%, 53%, 1);
-
-    background: linear-gradient(90deg, hsla(263, 54%, 53%, 1) 0%, hsla(236, 67%, 55%, 1) 100%);
-
-    background: -moz-linear-gradient(90deg, hsla(263, 54%, 53%, 1) 0%, hsla(236, 67%, 55%, 1) 100%);
-
-    background: -webkit-linear-gradient(90deg, hsla(263, 54%, 53%, 1) 0%, hsla(236, 67%, 55%, 1) 100%);
-
-    filter: progid: DXImageTransform.Microsoft.gradient(startColorstr="#7745C8", endColorstr="#3D47D9", GradientType=1);
-}
 
 /* .listHead{
     border-radius: 10px;
@@ -45,12 +29,6 @@
 
 .itemMsg {
     border-radius: 10px;
-    background-color: rgb(0, 0, 0);
-
-}
-
-.listItem {
-    list-style-type: none;
-    margin: 10px;
+    background-color: rgba(92, 90, 90, 0.651);
 
 }
