+++ client/src/components/admin/HelpUsMail/HelpUsMail.tsx
@@ -1,13 +1,16 @@
 import { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import st from './HelpUsMail.module.css'
+import s from '../../styles/styles.module.css'
 import { getReviews } from '../../../redux/actions/index'
+import ReviewCard from '../../Review/ReviewCard'
+import Bubble from '../../styles/bubbles'
 
 
 export default function HelpUsMail() {
 
     const dispatch: any = useDispatch()
-    
+
     useEffect(() => {
         dispatch(getReviews());
     }, [dispatch, getReviews]);
@@ -19,33 +22,10 @@ export default function HelpUsMail() {
     return (
 
         <div className={st.Container}>
-            <h1 className={st.title}> Feedback Mail  </h1>
-
-            <div className={st.listContainer}>
-                <ul className={st.listMessages}>
-                    {messages && messages.map((msg, index) => {
-                        return (
-                            <div className={st.divItem}>
-                                <div className={st.listHead}>
-                                    <li className={st.listItem}>
-                                        Review ID: {index + 1}
-                                    </li>
-                                    <li className={st.listItem}>
-                                        User: {msg.user}
-                                    </li>
-                                    <li className={st.listItem}>
-                                        Calification: {msg.calification}
-                                    </li>
-                                </div>
-                                <div className={st.itemMsg}>
-                                    <li className={st.listItem}>
-                                        Message: {msg.comment}
-                                    </li>
-                                </div>
-                            </div>
-                        )
-                    }).reverse()}
-                </ul>
+            <h3 className={st.title}> Feedback Mail  </h3>
+            <Bubble size='large' color='red' right='10%' />
+            <div className={st.listMessages}>
+                {messages && messages.map((item: review, index: number) => <ReviewCard username={item.user} rate={item.calification} message={item.comment} />).reverse()}
             </div>
         </div>
     )
