+++ client/src/components/admin/HomaPage/featuredInfo.module.css
@@ -1,53 +1,51 @@
-.featured{
+.featured {
     width: 100%;
     display: flex;
     justify-content: space-between;
 }
 
-.featuredItem{
+.featuredItem {
     flex: 1;
     margin: 0px 20px;
     padding: 30px;
     border-radius: 10px;
-    cursor: pointer;
-    -webkit-box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8); 
-    box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8);
+    box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.8);
 }
 
-.featuredTitle{
+.featuredTitle {
     font-size: 25px;
 }
 
-.featuredMoneyCont{
+.featuredMoneyCont {
     margin: 10px 0px;
     display: flex;
     align-items: center;
 }
 
-.featuredMoney{
+.featuredMoney {
     font-size: 30px;
     font-weight: 600;
 }
 
-.featuredMoneyRate{
+.featuredMoneyRate {
     display: flex;
     align-items: center;
     margin-left: 20px;
 }
 
-.featuredIcon{
+.featuredIcon {
     font-size: 25px !important;
     margin-left: 5px;
     color: green;
 }
 
-.featuredIcon_negative{
+.featuredIcon_negative {
     font-size: 25px !important;
     margin-left: 5px;
     color: red;
 }
 
-.featuredSub{
+.featuredSub {
     font-size: 17px;
     color: gray;
 }
