+++ client/src/components/admin/MainComponents/SideBar.module.css
@@ -1,19 +1,18 @@
 .sidebar {
 
-    border-top-left-radius: 0 !important;
-    border-bottom-left-radius: 0 !important;
-    display: grid;
-    grid-template-rows: 12% 55% 20% 10%;
-    position: sticky;
-    top: 95px;
-    height: fit-content;
-    width: 20vw;
-    background: rgba(219, 219, 219, 0.11);
-    box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-    backdrop-filter: blur(5.5px);
-    -webkit-backdrop-filter: blur(5.5px);
-    border-radius: 17px;
-    color: white;
+  border-top-left-radius: 0 !important;
+  border-bottom-left-radius: 0 !important;
+  display: grid;
+  grid-template-rows: 12% 55% 20% 10%;
+  position: sticky;
+  top: 95px;
+  height: fit-content;
+  width: 20vw;
+  background: rgba(219, 219, 219, 0.11);
+  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
+  backdrop-filter: blur(10px);
+  border-radius: 17px;
+  color: white;
 
 }
 
@@ -71,11 +70,9 @@ aside {
 .sidebarListItem.active,
 .sidebarListItem:hover {
   text-decoration: none;
-  background: linear-gradient(
-    45deg,
-    rgba(255, 255, 255, 0),
-    rgba(255, 255, 255, 0.4)
-  );
+  background: linear-gradient(45deg,
+      rgba(255, 255, 255, 0),
+      rgba(255, 255, 255, 0.4));
   box-shadow: 0 4px 30px rgba(100, 100, 100, 0.3);
 }
