+++ client/src/components/admin/MainComponents/TopBar.module.css
@@ -1,9 +1,9 @@
 .topbar {
-    margin-top: 1px;
+    margin-top: 1rem;
     width: 100%;
     height: 70px;
     position: sticky;
-    top: 0;
+    top: 1rem;
     background: rgba(219, 219, 219, 0.11);
     box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
     backdrop-filter: blur(5.5px);
@@ -42,7 +42,8 @@
     height: 4rem;
     margin-left: 1rem;
 }
-.imgEth{
+
+.imgEth {
     width: 50px;
     height: 50px;
 }
