+++ client/src/components/card/card.tsx
@@ -14,7 +14,7 @@ export default function Card({ name, id, image, current_price }: coinInfo) {
       <div className={style.img}>
         <img src={image} alt={name} />
       </div>
-      <h6 className={style.titulo} onClick={() => handleOnClick(id)}>
+      <h6 className={style.titulo}>
         {name}
       </h6>
